rename(donmain) : 패키지명 수정 및 패키지 추가
