+++ Titto_Backend/src/main/java/com/example/titto_backend/domain/Badge.java
@@ -1,0 +1,5 @@
+package com.example.titto_backend.domain;
+
+public enum Badge {
+  FIRST,SECOND,THIRD
+}
