+++ Titto_Backend/src/main/java/com/example/titto_backend/domain/Level.java
@@ -1,4 +1,4 @@
-package com.example.titto_backend.entity.dto;
+package com.example.titto_backend.domain;
 
 public enum Level {
   ONE,TWO,THREE,FOUR,FIVE
