--- Titto_Backend/src/main/java/com/example/titto_backend/entity/dto/Badge.java
@@ -1,5 +1,0 @@
-package com.example.titto_backend.entity.dto;
-
-public enum Badge {
-  FIRST,SECOND,THIRD
-}
